requirements.txt
pdfminer.six==20221105
spacy==3.7.2
deepface==0.0.79
opencv-python==4.8.0.76
librosa==0.10.1
soundfile==0.12.1
openai-whisper==20231117
streamlit==1.28.0
langchain==0.0.350
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
numpy==1.24.3
torch==2.1.0
sounddevice==0.4.6
wave==0.0.2 